Move to the single file `fftputils.py`.
